+++ src/Exceptions/AuthenticationException.php
@@ -2,9 +2,11 @@
 
 namespace Nuwave\Lighthouse\Exceptions;
 
+use GraphQL\Error\ClientAware;
+use GraphQL\Error\ProvidesExtensions;
 use Illuminate\Auth\AuthenticationException as IlluminateAuthenticationException;
 
-class AuthenticationException extends IlluminateAuthenticationException implements RendersErrorsExtensions
+class AuthenticationException extends IlluminateAuthenticationException implements ClientAware, ProvidesExtensions
 {
     public const MESSAGE = 'Unauthenticated.';
 
@@ -13,7 +15,7 @@ class AuthenticationException extends IlluminateAuthenticationException implemen
         return true;
     }
 
-    public function extensionsContent(): array
+    public function getExtensions(): array
     {
         return [
             'guards' => $this->guards,