+++ src/lighthouse.php
@@ -228,7 +228,6 @@ return [
 
     'error_handlers' => [
         \Nuwave\Lighthouse\Execution\ValidationErrorHandler::class,
-        \Nuwave\Lighthouse\Execution\ExtensionErrorHandler::class,
         \Nuwave\Lighthouse\Execution\ReportingErrorHandler::class,
     ],